Open notifications dialog

+++ app/src/main/java/projects/medicationtracker/Helpers/DBHelper.java
@@ -426,7 +426,7 @@ public class DBHelper extends SQLiteOpenHelper {
         String units = cursor.getString(cursor.getColumnIndexOrThrow(MED_UNITS));
         LocalDateTime startDate = TimeFormatting.stringToLocalDateTime(cursor.getString(
                 cursor.getColumnIndexOrThrow(START_DATE)));
-        long medId = cursor.getLong(cursor.getColumnIndexOrThrow(MED_ID));
+        int medId = cursor.getInt(cursor.getColumnIndexOrThrow(MED_ID));
         int frequency = cursor.getInt(cursor.getColumnIndexOrThrow(MED_FREQUENCY));
         int dosage = cursor.getInt(cursor.getColumnIndexOrThrow(MED_DOSAGE));
         String alias = cursor.getString(cursor.getColumnIndexOrThrow(ALIAS));
